Upgraded to VS2022

+++ lilia/llc_array_view.h
@@ -26,7 +26,7 @@ namespace llc
 		inline constexpr							array_view					(_tElement (&_dataElements)[_elementCount])									noexcept	: Data(_dataElements), Count(_elementCount)										{}
 
 		template <size_t _elementCount>
-		inline constexpr							array_view					(_tElement (&_dataElements)[_elementCount], uint32_t elementCount)						: Data(_dataElements), Count(::llc::min(_elementCount, elementCount))			{
+		inline constexpr							array_view					(_tElement (&_dataElements)[_elementCount], uint32_t elementCount)						: Data(_dataElements), Count(::llc::min((uint32_t)_elementCount, elementCount))			{
 			throw_if(elementCount > _elementCount, ::std::exception(""), "Element count out of range.");
 		}
